migrate off fzf; initial TUI draft
FZF is difficult to theme because, I think, it was never meant to be
fully customized by the user. Sure there is a config file and sure we
can use it, but I never managed to fix the alignment issues with the
config alone. It is also annoying having to shell out to a 3rd party
program that *might* or *might not* exist on the system. This commit
remedies that by deferring any menu rendering to
charmbracelet/bubbletea. It also gives us some nice colors, take that
ncurses!

Signed-off-by: NotAShelf <raf@notashelf.dev>
Change-Id: I6a6a6964693c845839305547aef88685fd1e4ea9